% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheduler.R
\name{create_emptyenv}
\alias{create_emptyenv}
\title{Create empty environment Function}
\usage{
create_emptyenv()
}
\description{
This function allows you to create empty environment.
}
\examples{
create_emptyenv()                       
}
\keyword{environment}
