% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheduler.R
\name{isEmptyQueue}
\alias{isEmptyQueue}
\title{A isEmptyQueue Function}
\usage{
isEmptyQueue()
}
\description{
This function allows you to check queue is empty or not.
}
\examples{
isEmptyQueue()    
}
\keyword{empty}
