% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheduler.R
\docType{data}
\name{Lsize}
\alias{Lsize}
\title{variables used in functions}
\format{An object of class \code{character} of length 1.}
\usage{
Lsize
}
\description{
variables used in functions
}
\keyword{datasets}
