% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheduler.R
\name{Node}
\alias{Node}
\title{Create Node Function}
\usage{
Node(val, node = NULL)
}
\description{
This function allows you to create node to be used in linked list/queue.
}
\examples{
Node() 
}
\keyword{node}
